# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hospital Management API",
        "description": "API for hospital management system",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/nurses/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "read specific nurse info",
                "description": "Read Nurse Endpoint.",
                "operationId": "readNurse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "nurse's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Object Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "update a nurse",
                "description": "Update Nurse Endpoint.",
                "operationId": "updateNurse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "nurse's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "$ref": "#/components/schemas/Gender"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "departement_id": {
                                        "type": "integer"
                                    },
                                    "specialization": {
                                        "$ref": "#/components/schemas/MedicalSpecialization"
                                    },
                                    "short_description": {
                                        "type": "string"
                                    },
                                    "rate": {
                                        "$ref": "#/components/schemas/Rate"
                                    },
                                    "doctor_id": {
                                        "type": "integer"
                                    },
                                    "assigned_at": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Object Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "delete a nurse",
                "description": "Delete Nurse Endpoint.",
                "operationId": "deleteNurse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "nurse's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Object Not Found"
                    }
                }
            }
        },
        "/api/nurses/": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "list all nurses",
                "description": "List Nurses Endpoint.",
                "operationId": "listNurses",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "operationId": "30f9eb106e755fd305297262461404e2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "phone_number",
                                    "address",
                                    "gender",
                                    "birth_date",
                                    "specialization",
                                    "short_description",
                                    "rate",
                                    "assinged_at",
                                    "departement_id"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "integer"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "specialization": {
                                        "type": "integer"
                                    },
                                    "short_description": {
                                        "type": "string"
                                    },
                                    "assigned_at": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "rate": {
                                        "type": "integer"
                                    },
                                    "departement_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Object Not Found"
                    }
                }
            }
        },
        "/api/nurses/me/": {
            "get": {
                "tags": [
                    "Nurse"
                ],
                "summary": "read the current nurse",
                "description": "Current Nurse Endpoint.",
                "operationId": "currentNurses",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/nurses/me": {
            "put": {
                "tags": [
                    "Nurse"
                ],
                "summary": "update personal info for current nurse",
                "description": "Update Nurse's Personal Info Endpoint.",
                "operationId": "updateCurrentNurse",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "$ref": "#/components/schemas/Gender"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Gender": {
                "title": "Gender Enum",
                "type": "integer",
                "enum": [
                    0,
                    1
                ]
            },
            "MedicalSpecialization": {
                "title": "Medical Specialization Enum",
                "type": "integer",
                "enum": [
                    0,
                    1,
                    2,
                    3,
                    4,
                    5
                ]
            },
            "Rate": {
                "title": "Rate Enum",
                "type": "integer",
                "enum": [
                    0,
                    1,
                    2,
                    3,
                    4
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}